requirements.txt
tree-sitter==0.21.3
setuptools
fastapi
uvicorn[standard]
PyYAML
python-dotenv

# LangChain (latest)
langchain
langchain-core
langchain-community
langchain-openai
langchain-chroma

# Vector DB + tokenization
chromadb
tiktoken
